[ADD] filters + query functionality for summary page
