+++ src/components/summary-page/index.ts
@@ -1,1 +1,3 @@
 export * from './components/summary-table';
+export * from './components/filters';
+export * from './components/hooks';
